requirements.txt
PyQt6
requests
pycryptodome 